Stop DHCP requests from launching ZTP tasks when hostname is changed.
Previous DHCP webhook tries to match the DHCP IP to a hostname in AWX
to prevent multiple ZTP workflows from being launched against a host.
If the hostname in AWX is changed, the DHCP webhook will create a new
host in AWX for the same device.  This is remedied here by embedding the
device's MAC address into the AWX description upon creation of a new AWX
host. New DHCP requests will create a new AWX host when no matching MAC
address is found.