image_management: Add 'Image-Flash-Copy-Version' template.
- 'Image-Flash-Copy' template renamed 'Image-Flash-Copy-File'.
- 'Image-Flash-Copy-Version' template added.

'Image-Flash-Copy-Version' template will ask for a software version and
either switch or router type image.  During a job run the switch will be
queried to get the model number and a image file will be automatically
selected to flash to that switch.  This will allow an inventory that
consists of multiple switch types to be updated to a specific version
with one job run.