awx_management: Fix localhost Python interpreter issue.
When the AWX server itself is the target for playbooks, there is an
issue with the Python Interpreter Discovery.  To be specific, the
awx_management templates in AWX have the 'Use Fact Cache' option
selected which will cache the discovered Python Interpreter in the host
facts.  Usually this is the 'platform-python' executable in the
task/execution environment container and this is not typically found on
the host system which will lead to errors when running a template that
need to establish a SSH connection to the host.  In order to remedy
this, one of the initial tasks is to save the host fact variable
'icx_awx_mgmt' which will contain all cached variables needed by the
'awx_management' playbooks.  The host facts are then cleared before an
SSH connection is established to the host system which will start a
Python discovery and then the playbook will gather facts on the host.
The discovered host Python executable path will then be saved in the
'icx_awx_mgmt' variable and this will be used for all further SSH tasks.
The localhost tasks can use the variable saved in
'ansible_playbook_python'.